refactor: split monolithic service into modules
